requirements.txt
streamlit==1.31.1
pillow==10.2.0
numpy==1.26.3
pyzbar==0.1.9 